pyproject.toml
[project]
name = "piclock"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "astral>=3.2",
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "metar>=2.0.1",
    "pillow>=10.4.0",
    "pygame-ce>=2.5.2",
    "sse-starlette>=3.3.4",
    "uvicorn[standard]>=0.44.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests"]
